+++ resources/aws_quicksight_theme.rb
@@ -22,54 +22,44 @@ property :aws_account_id, String,
          description: ""
 
 property :base_theme_id, String,
+         required: true,
          callbacks: {
            "base_theme_id is not a String" => lambda { |v| v.is_a? String },
-           "base_theme_id needs to be 1..2048 characters" => lambda { |v| v.length >= 1 && v.length <= 2048 },
+           "base_theme_id needs to be 1..512 characters" => lambda { |v| v.length >= 1 && v.length <= 512 },
            "base_theme_id must match pattern [\w\-]+" => lambda { |v| v =~ Regexp.new("/[\w\-]+/") },
          },
-         description: <<~'DESCRIPTION'
-           <p>The ID of the theme that a custom theme will inherit from. All themes inherit from one of
-      the starting themes defined by Amazon QuickSight. For a list of the starting themes, use
-      <code>ListThemes</code> or choose <b>Themes</b> from
-      within a QuickSight analysis. </p>
-         DESCRIPTION
+         description: ""
 
 property :configuration, Hash,
+         required: true,
          description: ""
 
 property :name, String,
          name_property: true,
+         required: true,
          callbacks: {
            "name is not a String" => lambda { |v| v.is_a? String },
            "name needs to be 1..2048 characters" => lambda { |v| v.length >= 1 && v.length <= 2048 },
          },
-         description: <<~'DESCRIPTION'
-           <p>A display name for the theme.</p>
-         DESCRIPTION
+         description: ""
 
 property :permissions, Array,
          callbacks: {
            "permissions is not a Array" => lambda { |v| v.is_a? Array },
          },
-         description: <<~'DESCRIPTION'
-           <p>A valid grouping of resource permissions to apply to the new theme.
-      </p>
-         DESCRIPTION
+         description: ""
 
 property :tags, Array,
          callbacks: {
            "tags is not a Array" => lambda { |v| v.is_a? Array },
          },
-         description: <<~'DESCRIPTION'
-           <p>A map of the key-value pairs for the resource tag or tags that you want to add to the
-      resource.</p>
-         DESCRIPTION
+         description: ""
 
 property :theme_id, String,
          required: true,
          callbacks: {
            "theme_id is not a String" => lambda { |v| v.is_a? String },
-           "theme_id needs to be 1..2048 characters" => lambda { |v| v.length >= 1 && v.length <= 2048 },
+           "theme_id needs to be 1..512 characters" => lambda { |v| v.length >= 1 && v.length <= 512 },
            "theme_id must match pattern [\w\-]+" => lambda { |v| v =~ Regexp.new("/[\w\-]+/") },
          },
          description: ""
@@ -79,12 +69,7 @@ property :version_description, String,
            "version_description is not a String" => lambda { |v| v.is_a? String },
            "version_description needs to be 1..512 characters" => lambda { |v| v.length >= 1 && v.length <= 512 },
          },
-         description: <<~'DESCRIPTION'
-           <p>A description of the first version of the theme that you're creating. Every time
-      <code>UpdateTheme</code> is called, a new version is created. Each version of the
-      theme has a description of the version in the <code>VersionDescription</code>
-      field.</p>
-         DESCRIPTION
+         description: ""
 
 # API URLs and mappings
 rest_api_collection "/AWS::QuickSight::Theme"